:recycle: リファクタリング
