+++ Otofuda-iOS/src/Main/Top/SearchGroup/SearchGroupVC+AVCaptureDelegate.swift
@@ -47,81 +47,7 @@ extension SearchGroupVC: AVCaptureMetadataOutputObjectsDelegate {
             
         }
         
-        
-        
-        
-        
-        
-//
-//            // QRコードのデータかどうかの確認
-//                if metadata.type == AVMetadataObject.ObjectType.qr {
-//
-//                    // 検出位置を取得
-//                    let barCode = videoLayer?.transformedMetadataObject(for: metadata) as! AVMetadataMachineReadableCodeObject
-//                    let box = CGRect(
-//                        x: barCode.bounds.minX,
-//                        y: barCode.bounds.minY + 250, // FIXME: 🐛このずれのせいでとりあえず250に設定している
-//                        width: barCode.bounds.width,
-//                        height: barCode.bounds.height
-//                    )
-//                    qrV!.frame = box
-//
-//                    if metadata.stringValue != nil {
-//                        // 検出データを取得
-//                        let strMetadata = metadata.stringValue!
-//
-//                        let roomID = cropURL(url: strMetadata)
-//
-//                        room = Room(name: roomID)
-//                        firebaseManager.observeSingle(path: room.url(), completion: { snapshot in
-//                            if let roomDict = snapshot.value as? [String: Any] {
-//                                self.enterRoom()
-//                                break cameraFor
-//                            }
-//                        })
-//                    }
-//                }
-//            }
-//        }
     }
     
-    func enterRoom(room: Room) {
-        if !isMatching {
-            let now = NSDate()
-            let formatter = DateFormatter()
-            formatter.dateFormat = "yyyy-M-d-H-m"
-            let current_date = formatter.string(from: now as Date)
-            
-            firebaseManager.post(path: room.url() + "date", value: current_date) // Roomがまだ未定義なのでエラーになる
-            
-            let uuid = UIDevice.current.identifierForVendor!.uuidString
-            
-            firebaseManager.observeSingle(path: room.url() + "member", completion: {(snapshot) in
-                var isExist: Bool = false
-                if let member = snapshot.value as? [String] {
-                    for user in member {
-                        if user == uuid {
-                            isExist = isExist || true
-                        }
-                    }
-                }
-                
-                if isExist {
-                    self.goNextVC(room: room)
-                    self.isMatching = true
-                } else {
-                    if var member = snapshot.value as? [String] {
-                        member.append( self.appDelegate.uuid )
-                        self.firebaseManager.post(path: room.url() + "member", value: member)
-                        self.goNextVC(room: room)
-                        self.isMatching = true
-                        return
-                    }
-                }
-            })
-            
-        }
-            
-            
-    }
+    
 }
